Refactoring: Narrow some value types (Species, Move, Form) (#3575)
In this pull request I've changed a ton of method signatures to reflect the more-narrow types of Species, Move# and Form; additionally, I've narrowed other large collections that stored lists of species / permitted values, and reworked them to be more performant with the latest API spaghetti that PKHeX provides. Roamer met locations, usually in a range of [max-min]<64, can be quickly checked using a bitflag operation on a UInt64. Other collections (like "Is this from Colosseum or XD") were eliminated -- shadow state is not transferred COLO<->XD, so having a Shadow ID or matching the met location from a gift/wild encounter is a sufficient check for "originated in XD".

+++ PKHeX.Core/Editing/Applicators/TechnicalRecordApplicator.cs
@@ -30,7 +30,7 @@ public static class TechnicalRecordApplicator
     /// </summary>
     /// <param name="pk">Pokémon to modify.</param>
     /// <param name="moves">Moves to set flags for. If a move is not a Technical Record, it is skipped.</param>
-    public static void SetRecordFlags(this ITechRecord8 pk, ReadOnlySpan<int> moves)
+    public static void SetRecordFlags(this ITechRecord8 pk, ReadOnlySpan<ushort> moves)
     {
         var permit = pk.TechRecordPermitFlags;
         var moveIDs = pk.TechRecordPermitIndexes;
