+++ PKHeX.Core/Legality/Encounters/EncounterTrade/EncounterTrade8.cs
@@ -19,7 +19,7 @@ public sealed record EncounterTrade8 : EncounterTrade, IDynamaxLevel, IRelearn, 
     public byte FlawlessIVCount { get; init; }
     public override Shiny Shiny { get; }
 
-    public EncounterTrade8(GameVersion game, int species, byte level, byte memory, ushort arg, byte feel, byte intensity, Shiny shiny = Shiny.Never) : base(game)
+    public EncounterTrade8(GameVersion game, ushort species, byte level, byte memory, ushort arg, byte feel, byte intensity, Shiny shiny = Shiny.Never) : base(game)
     {
         Species = species;
         Level = level;
